Move notify/alarms to modules
