+++ debian/changelog
@@ -1,3 +1,9 @@
+wb-rules (2.39.6) stable; urgency=medium
+
+  * Move notify/alarms to modules
+
+ -- Nikolay Korotkiy <nikolay.korotkiy@wirenboard.com>  Fri, 08 May 2026 17:00:00 +0400
+
 wb-rules (2.39.5) stable; urgency=medium
 
   * Update dependencies
