+++ wbrules/testrules_alarm.js
@@ -1,14 +1,12 @@
-// this stubs out the default alarm object
-global.__proto__.Notify = {
-  sendEmail: function sendEmail(to, subject, text) {
-    log('EMAIL TO: {} SUBJ: {} TEXT: {}', to, subject, text);
-  },
+// Override Notify used by modules loaded from lib.js.
+Notify.sendEmail = function (to, subject, text) {
+  log('EMAIL TO: {} SUBJ: {} TEXT: {}', to, subject, text);
+};
 
-  sendSMS: function sendSMS(to, text) {
-    log('SMS TO: {} TEXT: {}', to, text);
-  },
+Notify.sendSMS = function (to, text) {
+  log('SMS TO: {} TEXT: {}', to, text);
+};
 
-  sendTelegramMessage: function sendTelegramMessage(token, chatId, text) {
-    log("TELEGRAM MESSAGE TOKEN: {} CHATID: {} TEXT: {}", token, chatId, text);
-  },
+Notify.sendTelegramMessage = function (token, chatId, text) {
+  log("TELEGRAM MESSAGE TOKEN: {} CHATID: {} TEXT: {}", token, chatId, text);
 };
